% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_index_areas.R
\name{available_areas}
\alias{available_areas}
\title{Get available areas for index calculation}
\usage{
available_areas()
}
\value{
A character vector of available area names.
}
\description{
A helper function to see what areas are available in
\link{boundaries_data}. These area names can be passed to the \code{boundaries}
argument of \code{\link[=calc_index_areas]{calc_index_areas()}}.
}
\examples{
# See all available areas
available_areas()

# Use in calc_index_areas()
}
\seealso{
\itemize{
\item \code{\link[=calc_index_areas]{calc_index_areas()}}, calculate indices for specified areas
\item \link{boundaries_data}, the data object containing area boundaries
}
}
