% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_data}
\alias{boundaries_data}
\title{A list of northern and southern borders for U.S. West Coast areas}
\format{
\subsection{\code{boundaries_data}}{

A list with 10 entries, where each entry is a numeric vector of length two,
with the northern and southern boundaries for that area:
\describe{
\item{coastwide}{U.S. West Coast from Southern British Columbia to Northern
Mexico}
\item{WA}{Washington state}
\item{OR}{Oregon state}
\item{CA}{California state}
\item{}{North of Point Conception}
\item{}{South of Point Conception}
\item{}{North of Cape Mendocino}
\item{}{South of Cape Mendocino}
\item{}{North of Monterey Bay}
\item{}{South of Monterey Bay}
}
A list of named areas with a numeric vector of northern and southern
boundaries for each area. Current areas include
\itemize{
\item coastwide for the U.S. West Coast
\item Washington (WA)
\item Oregon (OR)
\item California (CA)
\item North of Point Conception
\item South of Point Conception
\item North of Cape Mendocino
\item South of Cape Mendocino
\item North of Monterey Bay
\item South of Monterey Bay
}
}
}
\usage{
boundaries_data
}
\description{
Northern and southern borders for areas of interest along the U.S. West
Coast. Where, some areas are specific to management regions of the Pacific
Fisheries Management Council and some are entirely related to ecological
boundaries from currents, etc. that limit movement of populations.
}
\keyword{datasets}
