% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{california_current_grid}
\alias{california_current_grid}
\title{California Current prediction grid}
\format{
A data frame
}
\source{
Keller et al. 2017
}
\usage{
california_current_grid
}
\description{
A 4km prediction grid for the West Coast Groundfish Bottom Trawl Survey,
used to generate biomass indices from predictions.
}
\keyword{datasets}
