% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{configuration}
\alias{configuration}
\title{Configuration data for west coast index standardization models}
\format{
A data frame
}
\usage{
configuration
}
\description{
Configuration settings for different species and model specifications.
}
\keyword{datasets}
