% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_grid.R
\name{filter_grid}
\alias{filter_grid}
\title{Filter a grid created with \code{\link[=lookup_grid]{lookup_grid()}}}
\usage{
filter_grid(grid, boundary_north, boundary_south)
}
\arguments{
\item{grid}{A grid returned from \code{\link[=lookup_grid]{lookup_grid()}}.}

\item{boundary_north, boundary_south}{Real numbers providing the northern and
southern latitudinal boundaries for your desired grid.}
}
\value{
A data frame with the same structure as \code{grid}.
}
\description{
Filter a grid created with \code{\link[=lookup_grid]{lookup_grid()}}
}
\author{
Kelli F. Johnson
}
