% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_common_name.R
\name{format_common_name}
\alias{format_common_name}
\title{Format common name for directory management}
\usage{
format_common_name(x)
}
\description{
Format common name for directory management
}
\author{
Chantel R. Wetzel
}
