% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_strata.R
\name{format_strata}
\alias{format_strata}
\title{Format the strata for a species distribution model}
\usage{
format_strata(
  strata = nwfscSurvey::GetStrata.fn("coast"),
  min_depth = 55,
  max_depth = 1280
)
}
\arguments{
\item{strata}{An object returned from \code{\link[nwfscSurvey:GetStrata.fn]{nwfscSurvey::GetStrata.fn()}}.}

\item{min_depth, max_depth}{The minimum and maximum depths you want to use
for all strata if you want to override the values in \code{strata}.}
}
\description{
Format the strata for a species distribution model
}
\examples{
format_strata()
format_strata(nwfscSurvey::GetStrata.fn("sablefish"))
}
