% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diagnostics.R
\name{get_diagnostics}
\alias{get_diagnostics}
\title{Function to retreive model estimates and plot diagnositcs for sdmTMB
index of abundance model runs.}
\usage{
get_diagnostics(dir, fit, prediction_grid)
}
\arguments{
\item{dir}{A directory location where results and figures will be saved.}

\item{fit}{List created by the sdmTMB:: fit function.}

\item{prediction_grid}{The prediction grid for the survey that sdmTMB will use
to make model predictions to.}
}
\description{
Function to retreive model estimates and plot diagnositcs for sdmTMB
index of abundance model runs.
}
\examples{
\dontrun{
	get_diagnostics(
		sppdir = 'directory location for species specific run', 
		fit = fit, # object created by sdmTMB::fit function 
		prediction_grid = year, 
		survey = survey, # e.g. NWFSC.Combo 
		obs = 'lognormal')
}




}
\author{
Chantel Wetzel
}
