% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_grid.R
\name{lookup_grid}
\alias{lookup_grid}
\title{Lookup the prediction grid}
\usage{
lookup_grid(
  x,
  years,
  max_latitude,
  min_latitude,
  max_longitude,
  min_longitude,
  mean_depth,
  sd_depth,
  max_depth = Inf,
  data = california_current_grid
)
}
\arguments{
\item{x}{The raw survey data}

\item{years}{The years to use for predictions}

\item{max_latitude}{The maximum latitude for predictions}

\item{min_latitude}{The minimum latitude for predictions}

\item{max_longitude}{The maximum longitude for predictions}

\item{min_longitude}{The minimum longitude for predictions}

\item{mean_depth}{The mean depth to use}

\item{sd_depth}{The sd of depth to use}

\item{max_depth}{Maximum depth, defaults to Inf}

\item{data}{the name of the grid, defaults to california_current_grid (WCGBTS)}
}
\description{
Lookup the prediction grid
}
