% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_base.R
\name{map_base}
\alias{map_base}
\title{Plot a map to use as the base for other maps}
\usage{
map_base(x_min = -155, x_max = -115, y_min = 20, y_max = 48)
}
\arguments{
\item{x_min, x_max, y_min, y_max}{Minimum and maximum values for the extent of
the desired map. These need to be in decimal degrees. The default is to
map the California Current from the U.S./Canada border to the
U.S./Mexico border.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Plot a map with lines for the United States, including islands.
}
\examples{
\dontrun{
# Return a blank base map
map_base()
# TODO: example with map_base second
}
}
\author{
Kelli F. Johnson
}
