% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_density.R
\name{map_density}
\alias{map_density}
\title{Plot the predicted biomass density by year.}
\usage{
map_density(
  predictions,
  save_prefix = fs::path(getwd(), "density"),
  n_row = 1,
  n_col = 2,
  column_grep = "^est[2]*$",
  tile_size = c(2000, 2000)
)
}
\arguments{
\item{predictions}{Output from \code{\link[stats:predict]{stats::predict()}}.}

\item{save_prefix}{A string providing the desired prefix to use when saving
the density plots. The default is to save them in the current working
directory using the name \code{"density[0-9]{2}.png"}. Obviously, the same
outcome would result from just \code{"density"} but the use of \code{fs::path()}
in the default is present to provide an example for users who want to
change the path to something other than their current directory.}

\item{n_row, n_col}{Integers specifying the number of rows and columns for
the faceted figures. The default is one row and two columns.}

\item{column_grep}{A regular expression string that will be used to search
the column names of \code{predictions} to find the column name that will be
passed to \code{ggplot2::geom_tile(ggplot2::aes(fill = ))}. The default results
in either \code{"est"} or \code{"est2"} depending on if the model is a delta model
or not.}

\item{tile_size}{A vector of length two with numeric values specifying the
tile width and height that will be passed to
\code{ggplot2::geom_tile(width, height)}.`}
}
\value{
A \pkg{ggplot2} \pkg{ggforce} object is returned.
}
\description{
Plot the predicted biomass density by year.
}
\seealso{
\itemize{
\item \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}} for how the raster is created
\item \code{\link[ggforce:facet_wrap_paginate]{ggforce::facet_wrap_paginate()}} for how the pages are created
\item \code{\link[ggforce:n_pages]{ggforce::n_pages()}} for how many pages the object has
}
}
\author{
Chantel R. Wetzel
}
