% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_effects.R
\name{map_effects}
\alias{map_effects}
\title{Plot the fixed and spatial effects from the fit predictions}
\usage{
map_effects(predictions, dir, verbose = FALSE)
}
\arguments{
\item{predictions}{A data frame of predictions made by sdmTMB using the
predict function based on the \code{newdata} prediction grid passed to the
spatial, and spatio-temporal based on the fit model estimates.}

\item{dir}{A directory location where results and figures will be saved.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{FALSE}.}
}
\description{
Plot the fixed and spatial effects from the fit predictions
}
\author{
Chantel R. Wetzel
}
