% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_year_re.R
\name{map_year_re}
\alias{map_year_re}
\title{Plot the annual random effects predictions by year
for the fit model}
\usage{
map_year_re(predictions, dir, nrow = 3, ncol = 4, verbose = FALSE)
}
\arguments{
\item{predictions}{A data frame of predictions made by sdmTMB using the
predict function based on the \code{newdata} prediction grid passed to the
spatial, and spatio-temporal based on the fit model estimates.}

\item{dir}{A directory location where results and figures will be saved.}

\item{nrow}{The number of rows to plot for multi-year figures.}

\item{ncol}{The number of columns to plot for multi-year figures.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{FALSE}.}
}
\description{
Plot the annual random effects predictions by year
for the fit model
}
\examples{
\dontrun{
	predictions <- predict(
	    fit , 
	    newdata = prediction_grid
	) 
	predictions[, c("lon", "lat")] <- round(predictions[, c("Lon", "Lat")], 1)
	map_year_re(
		predictions = predictions, 
		dir = file.path(sppdir, survey, "index", obs)
	)
 }


}
\author{
Chantel R. Wetzel
}
