% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapdata}
\alias{mapdata}
\title{Mapping data from rnaturalearth}
\format{
Class is sf and data.frame
}
\usage{
mapdata
}
\description{
This data object represents the output of a call to rnaturalearth using
mapdata <- rnaturalearth::ne_countries(scale = "medium",
returnclass = "sf", country = "united states of america") and is included
here to avoid dependencies on rnaturalearth
}
\keyword{datasets}
