% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fixed_effects_para.R
\name{plot_fixed_effects_para}
\alias{plot_fixed_effects_para}
\title{Plot the annual parameter estimates from the fit model}
\usage{
plot_fixed_effects_para(data, dir)
}
\arguments{
\item{data}{List created by the sdmTMB:: fit function}

\item{dir}{A directory location where results and figures will be saved.}
}
\description{
Plot the annual parameter estimates from the fit model
}
\examples{
\dontrun{
 plot_fixed_effects_para(
       data = fit, 
       dir = file.path(sppdir, survey, "index", obs))
}


}
\author{
Chantel Wetzel
}
