% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indices.R
\name{plot_indices}
\alias{plot_indices}
\title{Plot the indices}
\usage{
plot_indices(
  data,
  save_loc = getwd(),
  file_name = "index.png",
  legend_loc = "right"
)
}
\arguments{
\item{data}{A data frame created by \code{\link[sdmTMB:get_index]{sdmTMB::get_index()}} with an additional
column giving the \code{.id} of each index called \code{area}. If \code{area} is not
not present, then all data points will be assumed to be from one default
area. It behooves you to make \code{area} an ordered factor such that the areas
are plotted in the order that you desire.}

\item{save_loc}{A string providing the location of the directory you wish to
save the png file to. The default is your current working directory, and if
\code{NULL} no directory will be created}

\item{file_name}{A string giving the file name. The default is
\code{"index.png"}. If NULL, the file is not written but the ggplot object is returned}

\item{legend_loc}{Location for the legend to be added to the figure where
the options are \code{"top"}, \code{"bottom"}, \code{"left"}, \code{"right"}, or a vector of
length 2 giving relative coordinate positions. The default is \code{"right"}.}
}
\value{
Creates and saves a plot of the abundance indices in the directory location
specified by the save_loc function input.
}
\description{
Plot the indices
}
\details{
TODO
\itemize{
\item Think about adding some plotting information back in to have a title
on the figure.
}
}
\author{
Chantel R. Wetzel
}
