% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_density.R
\name{plot_map_density}
\alias{plot_map_density}
\title{Plot the predicted biomass density by year.}
\usage{
plot_map_density(predictions, dir, ncol = 4, nrow = 3, verbose = FALSE)
}
\arguments{
\item{predictions}{A data frame of predictions made by sdmTMB using the
predict function based on the \code{newdata} prediction grid passed to the
spatial, and spatio-temporal based on the fit model estimates.}

\item{dir}{A directory location where results and figures will be saved.}

\item{ncol}{The number of columns to plot for multi-year figures.}

\item{nrow}{The number of rows to plot for multi-year figures.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{FALSE}.}
}
\description{
Plot the predicted biomass density by year.
}
\author{
Chantel R. Wetzel
}
