% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mesh.R
\name{plot_mesh}
\alias{plot_mesh}
\title{Plot a mesh object}
\usage{
plot_mesh(mesh, file_name = NULL)
}
\arguments{
\item{mesh}{A mesh object from \code{\link[sdmTMB:make_mesh]{sdmTMB::make_mesh()}}}

\item{file_name}{Optional file path to save the plot. If \code{NULL}, plot is
returned but not saved.}
}
\value{
A ggplot object
}
\description{
Plot a mesh object
}
\examples{
\dontrun{
mesh <- sdmTMB::make_mesh(data, xy_cols = c("x", "y"), n_knots = 500)

# Return plot without saving
p <- plot_mesh(mesh, file_name = NULL)
print(p)
}
}
