% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot the residuals from a model fit by \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}}
\usage{
plot_residuals(data, dir, nrow = 3, ncol = 4)
}
\arguments{
\item{data}{A list returned from \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.}

\item{dir}{A directory location where results and figures will be saved.}

\item{nrow}{The number of rows to plot for multi-year figures.}

\item{ncol}{The number of columns to plot for multi-year figures.}
}
\description{
Plot the residuals from a model fit by \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}
}
\author{
Chantel R. Wetzel
}
