% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_and_format_data.R
\name{pull_and_format_data}
\alias{pull_and_format_data}
\title{Pull and format \code{data} for use in downstream functions.}
\usage{
pull_and_format_data(
  configuration_to_run,
  common_name = NULL,
  survey = "NWFSC.Combo",
  years = 2003:2050,
  verbose = TRUE
)
}
\arguments{
\item{configuration_to_run}{A data frame based on the configuration data object
that is used to pull and format data list used in downstream functions (e.g.,
\code{\link[=run_sdmtmb]{run_sdmtmb()}}). If a configuration_to_run is specified, the function will
pull data from \code{\link[nwfscSurvey:pull_catch]{nwfscSurvey::pull_catch()}} for each line and will create a
data list that is used by \code{\link[=run_sdmtmb]{run_sdmtmb()}}.}

\item{common_name}{A character string used to pull data from \code{\link[nwfscSurvey:pull_catch]{nwfscSurvey::pull_catch()}}.
This is only used if \code{configuration_to_run} is not specified.}

\item{survey}{A character string used to pull data from \code{\link[nwfscSurvey:pull_catch]{nwfscSurvey::pull_catch()}}.
Commonly used values are \code{NWFSC.Combo}, \code{Triennial}, \code{AFSC.Slope}, and \code{NWFSC.Slope}.
This is only used if \code{configuration_to_run} is not specified.}

\item{years}{A numeric vector used to pull data from \code{\link[nwfscSurvey:pull_catch]{nwfscSurvey::pull_catch()}}.
This is only used if \code{configuration_to_run} is not specified.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{FALSE}.}
}
\value{
A list of configurations and data
\itemize{
\item \code{species}
\item \code{fxn}
\item \code{source}
\item \code{family}
\item \code{formula}
\item \code{min_depth}
\item \code{max_depth}
\item \code{min_latitude}
\item \code{min_year}
\item \code{max_year}
\item \code{anisotropy}
\item \code{knots}
\item \code{spatiotemporal1}
\item \code{spatiotemporal2}
\item \code{share_range}
\item \code{used}
\item \code{data_raw}
\item \code{data_filtred}
}
}
\description{
A list is returned with specifications to run a sdmTMB model, formatted and
filtered data, and a prediction grid used to estimate indices of abundance
in \code{\link[=calc_index_areas]{calc_index_areas()}} and \code{\link[=diagnose]{diagnose()}}.
}
\author{
Chantel R. Wetzel and Kelli F. Johnson
}
