% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_data.r
\name{sanity_data}
\alias{sanity_data}
\title{Creates a data frame of information from \code{\link[sdmTMB:sanity]{sdmTMB::sanity()}}}
\usage{
sanity_data(fit)
}
\arguments{
\item{fit}{A returned object from \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.}
}
\value{
A csv with three columns, \code{name}, \code{logical}, and \code{text}, containing the
name of the test, a logical for the results of the test, and a text
explanation of what the test really means.
}
\description{
The returned data frame is useful for model checking after running several
models and not having access to the R console while running the models.
}
\seealso{
\itemize{
\item \code{\link[sdmTMB:sanity]{sdmTMB::sanity()}}
\item \code{\link[=diagnose]{diagnose()}}
}
}
\author{
Kelli F. Johnson
}
