% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_default_text.R
\name{write_default_text}
\alias{write_default_text}
\title{Write default text describing the index standardizations run from the {indexwc} package}
\usage{
write_default_text(species_list = NULL)
}
\arguments{
\item{species_list}{vector with strings matching a subset of those in the "species" column of the configuration file. If NULL, this will be the set of species that have at least one TRUE value in the "used" column.}
}
\description{
Writes both a markdown (.md) file which can be copied into an Quarto assessment report as
well as the PDF showing the formatted text. Relies on \code{inst/default_text.Rmd} and
\code{inst/child.Rmd} files where the child file is populated with information for each survey
associated with the species.
}
\examples{
\dontrun{
write_default_text()
}
}
\author{
Ian G. Taylor
}
