% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hessian.R
\name{write_hessian}
\alias{write_hessian}
\title{Write a file with a single logical regarding the hessian status}
\usage{
write_hessian(fit, file = "hess_logical.txt")
}
\arguments{
\item{fit}{A model object returned from \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.}

\item{file}{A path to the file you want to write. The default path is
\code{hess_logical.txt}, which will be saved in your current working directory.}
}
\value{
A logical is invisibly returned providing the status of the hessian.
}
\description{
Write a file with a single logical regarding the hessian status
}
\author{
Kelli F. Johnson
}
