% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yellowtail}
\alias{yellowtail}
\title{Yellowtail rockfish survey data}
\format{
A data frame
}
\usage{
yellowtail
}
\description{
West Coast Groundfish Bottom Trawl Survey data for yellowtail rockfish, 2003 - 2023. Data
represents a subset of all hauls in each year, within the 55m - 425m depth range. The
latitudinal extent is 33.5 - 48.5 degrees N.
}
\keyword{datasets}
