% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_index_areas.R
\name{calc_index_areas}
\alias{calc_index_areas}
\title{Calculate indices by areas, e.g., Washington, Oregon, and California}
\usage{
calc_index_areas(data, fit, prediction_grid, dir)
}
\arguments{
\item{data}{A data frame containing tow-level information on catch weight
and effort. Additional columns can be present such as depth.}

\item{fit}{List created by the sdmTMB:: fit function.}

\item{prediction_grid}{The prediction grid for the survey that sdmTMB will
use to make model predictions to.}

\item{dir}{A directory location where results and figures will be saved.}
}
\value{
A {ggplot2} object is returned with a figure of the indices. This figure is
also saved to the disk. But, the main benefit of returning the figure is
that users can change parts of the figure and re-save it if they wish and
the calculated indices will be available in \verb{*[["data"]]} where you should
replace \code{*} with your saved object name.
}
\description{
Calculate indices by areas, e.g., Washington, Oregon, and California
}
\author{
Kelli F. Johnson
}
