% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{Retrieve model estimates and plot diagnostics}
\usage{
diagnose(dir, fit, prediction_grid)
}
\arguments{
\item{dir}{A directory location where results and figures will be saved.}

\item{fit}{List created by the sdmTMB:: fit function.}

\item{prediction_grid}{The prediction grid for the survey that sdmTMB will
use to make model predictions to.}
}
\description{
Retrieve model estimates and plot diagnostics
}
\details{
Parameter definitions:
\itemize{
\item Range: A derived parameter that defines the distance at which two
points are effectively independent (actually about 13\% correlated).
If the share_range argument is changed to FALSE then the spatial
and spatiotemporal ranges will be unique, otherwise the default is
for both to share the same range.
\item Phi: Observation error scale parameter (e.g., SD in Gaussian).
\item sigma_O: Standard deviation of the spatial process ("Omega").
\item sigma_E: Standard deviation of the spatiotemporal process ("Epsilon").
}
}
\author{
Chantel R. Wetzel
}
