% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Format data pull from the data warehouse}
\usage{
format_data(data)
}
\arguments{
\item{data}{Catch data pulled from the data warehouse}
}
\value{
A dataframe of formatted catch data.
}
\description{
Format data pull from the data warehouse
}
\examples{
formatted_data <- format_data(data = catch_data)

}
\author{
Chantel Wetzel
}
