% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_results.R
\name{get_area_results}
\alias{get_area_results}
\title{Calculate index based upon the selected areas based on the prediction grid}
\usage{
get_area_results(
  area_name,
  area_bounds,
  fit,
  full_grid,
  calculate_cog = FALSE,
  bias_correct = TRUE
)
}
\arguments{
\item{area_name}{A character vector specifying which areas to calculate
indices for. These should be names from \link{boundaries_data}. The default is
\code{"Coastwide"}, which calculates only the Coastwide index}

\item{area_bounds}{Prediction grid as modified by \code{\link[=filter_boundaries]{filter_boundaries()}} and
\code{boundaries}.}

\item{fit}{A fitted sdmTMB model object returned by \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.}

\item{full_grid}{Filtered grid created by \code{\link[=filter_grid]{filter_grid()}}, passed to
\code{\link[sdmTMB:get_index]{sdmTMB::get_index()}} as \code{newdata}.}

\item{calculate_cog}{Logical. If \code{TRUE}, center of gravity estimates will also be
calculated using \code{\link[sdmTMB:get_cog]{sdmTMB::get_cog()}} for each area. Defaults to \code{FALSE}.
Note that COG results are returned in long format with separate rows for
the X (easting) and Y (northing) coordinate axes.}

\item{bias_correct}{Logical. If \code{TRUE} \code{\link[sdmTMB:get_index]{sdmTMB::get_index()}} and
\code{\link[sdmTMB:get_cog]{sdmTMB::get_cog()}} will use bias correction to account for the non-linear
transformation of random effects when calculating the index. Will be faster
if set to \code{FALSE}, but is \code{TRUE} by default}
}
\description{
Calculate index based upon the selected areas based on the prediction grid
}
\details{
Add details
}
\author{
Kelli F. Johnson, Chantel Wetzel, and Eric Ward
}
