% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_residuals.R
\name{map_residuals}
\alias{map_residuals}
\title{Plot the data from a model fit by \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}}
\usage{
map_residuals(data, n_row = 1, n_col = 2, save_prefix = NULL)
}
\arguments{
\item{data}{A data frame with at least the following three columns:
\itemize{
\item \code{X},
\item \code{Y}, and
\item \code{data},
where the data column is the result of \code{\link[stats:residuals]{stats::residuals()}}.
}}

\item{n_row, n_col}{Integers specifying the number of rows and columns for
the faceted figures. The default is one row and two columns.}

\item{save_prefix}{A string providing the desired prefix to use when saving
the density plots. The default is to save them in the current working
directory using the name \code{"density[0-9]{2}.png"}. Obviously, the same
outcome would result from just \code{"density"} but the use of \code{file.path()}
in the default is present to provide an example for users who want to
change the path to something other than their current directory.}
}
\description{
Plot the data from a model fit by \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}
}
\author{
Chantel R. Wetzel
}
