% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nwfsc.R
\name{plot_map}
\alias{plot_map}
\title{Map using \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}}
\usage{
plot_map(data, column)
}
\arguments{
\item{data}{Predictions data frame created by \code{\link[stats:predict]{stats::predict()}} and the
prediction grid.}

\item{column}{The column name to plot within the plotting function (e.g.,
est_non_rf2, omega_s2, epsilon_st2).}
}
\description{
Map using \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\author{
Chantel R. Wetzel
}
