% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{Plot the QQ and 1:1 line}
\usage{
plot_qq(data, dir)
}
\arguments{
\item{data}{List created by the sdmTMB:: fit function}

\item{dir}{A directory location where results and figures will be saved.}
}
\description{
Plot the QQ and 1:1 line
}
\author{
Chantel Wetzel
}
