% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{Plot the quantile-quantile for \code{fit} with a 1:1 line}
\usage{
plot_qq(data, dir)
}
\arguments{
\item{data}{A list created by \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.}

\item{dir}{A directory location where results and figures will be saved.}
}
\description{
Plot the quantile-quantile for \code{fit} with a 1:1 line
}
\author{
Chantel R. Wetzel
}
