% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Estimate indices of abundance for combinations of surveys and populations}
\usage{
run(
  data = nwfscSurvey::pull_catch(survey = "NWFSC.Combo"),
  family = gaussian(link = "identity"),
  dir_main = getwd()
)
}
\arguments{
\item{data}{A data frame with the following columns:
\itemize{
\item survey_name
\item common_name
\item 
}

Or a data frame pulled from the NWFSC data warehouse. See the function
call for the default, which is to pull data for \emph{EVERY} species present
in the West Coast Groundfish Bottom Trawl Survey.}

\item{dir_main}{A string specifying the file path to the working directory
that will be populated with additional directories where the results will
be saved.}
}
\description{
Run {sdmTMB} to estimate indices of abundance for combinations of surveys
and populations present in \code{data}.
}
\concept{run}
