% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{A deprecated function, use \code{\link[=run_sdmtmb]{run_sdmtmb()}}}
\usage{
run(data, family, formula, dir_main = getwd(), n_knots = 500, ...)
}
\arguments{
\item{data}{The data object being passed in}

\item{family}{The name of the family used for the response variable}

\item{formula}{The formula used in the model}

\item{dir_main}{The name of the main directory}

\item{n_knots}{The number of knots to specify for the SPDE mesh}

\item{...}{Extra arguments}
}
\value{
A \code{list} of {sdmTMB} \code{list}s, where each element in the list is the returned
object from \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}} when fitting data to a model and of the class
\code{sdmTMB}. The list of lists is because \code{\link[purrr:map2]{purrr::map2()}} is used to split the
input data by species and survey/source in case your data contain
combinations of these two categories in a long data frame.
}
\description{
Split your data into a list of data frames, e.g.,
\code{dplyr::split(data, common_name)} and then use \code{\link[purrr:map]{purrr::map()}} on the
resulting list with \code{.f = run_sdmtmb}.
}
\concept{run}
