% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sdmtmb.R
\name{run_sdmtmb}
\alias{run_sdmtmb}
\title{Run \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}}
\usage{
run_sdmtmb(
  dir_main = getwd(),
  data,
  family,
  formula,
  n_knots = 500,
  share_range = FALSE,
  sdmtmb_control = sdmTMB::sdmTMBcontrol(newton_loops = 3),
  ...
)
}
\arguments{
\item{dir_main}{A string specifying a path where results will be saved. The
default is your current working directory.}

\item{data}{A data frame containing tow-level information on catch weight
and effort. Additional columns can be present such as depth.}

\item{family}{The family and link. Supports \code{\link[=gaussian]{gaussian()}}, \code{\link[=Gamma]{Gamma()}},
\code{\link[=binomial]{binomial()}}, \code{\link[=poisson]{poisson()}}, \code{\link[sdmTMB:families]{Beta()}},
\code{\link[sdmTMB:families]{nbinom2()}},
\code{\link[sdmTMB:families]{truncated_nbinom2()}},
\code{\link[sdmTMB:families]{nbinom1()}},
\code{\link[sdmTMB:families]{truncated_nbinom1()}},
\code{\link[sdmTMB:families]{censored_poisson()}},
\code{\link[sdmTMB:families]{gamma_mix()}},
\code{\link[sdmTMB:families]{lognormal_mix()}},
\code{\link[sdmTMB:families]{student()}},
\code{\link[sdmTMB:families]{tweedie()}}, and
\code{\link[sdmTMB:families]{gengamma()}}.
Supports the delta/hurdle models:
\code{\link[sdmTMB:families]{delta_beta()}},
\code{\link[sdmTMB:families]{delta_gamma()}},
\code{\link[sdmTMB:families]{delta_gamma_mix()}},
\code{\link[sdmTMB:families]{delta_lognormal_mix()}},
\code{\link[sdmTMB:families]{delta_lognormal()}}, and
\code{\link[sdmTMB:families]{delta_truncated_nbinom2()}},
For binomial family options, see 'Binomial families' in the Details
section below.}

\item{formula}{Model formula. IID random intercepts and slopes are possible using
\pkg{lme4} syntax, e.g., \code{+ (1 | g)} or \code{+ (0 + depth | g)} or \code{+ (1 + depth | g)} where \code{g} is a column of class
character or factor representing groups. Penalized splines are possible via
\pkg{mgcv} with \code{s()}. Optionally a list for delta (hurdle) models.  See
examples and details below.}

\item{n_knots}{An integer specifying the number of knots you want in your
mesh that is created by {fmesher}. More knots is not always better. The
default is to use 500 knots. Future work will look at specifying a
threshold distance between points rather than number of knots.}

\item{share_range}{Logical, whether or not to share the range between the
spatial and spatiotemporal fields. This defaults to FALSE, but adds extra
parameters. The default in sdmTMB is TRUE, and sharing the range may improve
estimation for data limited applications}

\item{sdmtmb_control}{Optional list, in the format of \code{\link[sdmTMB:sdmTMBcontrol]{sdmTMB::sdmTMBcontrol()}}.
By default, this is includes 3 newton loops}

\item{...}{Optional arguments passed to \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.}
}
\value{
An list object of class sdmTMB. This object is returned by \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}
when fitting data to the model.
}
\description{
Run \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}
}
\author{
Chantel R. Wetzel
}
